+++ lib/iListenAttentively.d.ts
@@ -1,7 +1,7 @@
 /** 监听ID */
-type ListenerId = Number;
+type ListenerId = number;
 /** 事件名 */
-type EventName = String;
+type EventName = string;
 
 /** 事件优先级 */
 export enum EventPriority {
@@ -18,7 +18,7 @@ export enum EventPriority {
 };
 
 /** 获取自身插件名 */
-export function getPluginName(): String;
+export function getPluginName(): string;
 
 /** 获取所有事件别名 */
 export function getAllEventAlias(): Record<EventName, EventName>;
@@ -38,34 +38,34 @@ export function getEventName(
 /** 通过维度名字获取ID */
 export function getDimensionNameFromId(
     /** 维度ID */
-    dimensionId: Number
-): String | undefined;
+    dimensionId: number
+): string | undefined;
 
 /** 通过维度ID获取名字 */
 export function getDimensionIdFromName(
     /** 维度名字 */
-    dimensionName: String
-): Number | undefined;
+    dimensionName: string
+): number | undefined;
 
 /** 获取所有已注册事件 */
 export function getAllEvent(): {
     /** 事件名 */
     eventName: EventName,
     /** 模组名 */
-    modName: String
+    modName: string
 }[];
 
 /** 获取指定插件的已注册事件 */
 export function getAllEvent(
     /** 模组名 */
-    modName: String
+    modName: string
 ): EventName[];
 
 /** 获取事件监听数量 */
 export function getListenerCount(
     /** 事件名 */
     eventName: EventName
-): Number;
+): number;
 
 /** 删除事件监听 */
 export function removeListener(
@@ -113,14 +113,14 @@ export function publish(
     /** 事件数据 */
     eventData: NbtCompound,
     /** 模组名 */
-    modName: String
+    modName: string
 ): void;
 
 interface EventListenerData extends NbtCompound extends NbtList {
-    [key: String | Number]: EventListenerData | undefined;
-    [Symbol.iterator]: () => Iterator<[String | Number, EventListenerData]>;
+    [key: string | number]: EventListenerData | undefined;
+    [Symbol.iterator]: () => Iterator<[string | number, EventListenerData]>;
 
-    toSNBT(index = -1): String;
+    toSNBT(index = -1): string;
 }
 
 /** 安装事件监听(需事件注册后安装) */
@@ -135,170 +135,170 @@ export function emplaceListener(
     /** 事件优先级 */
     priority: EventPriority.Normal = EventPriority.Normal,
     /** 模组名 */
-    modName: String | undefined = getPluginName()
+    modName: string | undefined = getPluginName()
 ): ListenerId;
 
 /** 根据内存地址/序列化信息获取玩家 */
 export function getPlayer(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
+    info: number | NbtCompound | EventListenerData
 ): Player | undefined;
 
 /** 根据内存地址/序列化信息获取实体 */
 export function getActor(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
+    info: number | NbtCompound | EventListenerData
 ): Entity | undefined;
 
 /** 根据内存地址/序列化信息获取物品 */
 export function getItemStack(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
+    info: number | NbtCompound | EventListenerData
 ): Item | undefined;
 
 /** 根据内存地址/序列化信息获取方块 */
 export function getBlock(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
+    info: number | NbtCompound | EventListenerData
 ): Block | undefined;
 
 /** 根据内存地址/序列化信息获取方块实体 */
 export function getBlockActor(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
+    info: number | NbtCompound | EventListenerData
 ): BlockEntity | undefined;
 
 /** 根据内存地址/序列化信息获取容器 */
 export function getContainer(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
+    info: number | NbtCompound | EventListenerData
 ): Container | undefined;
 
 /** 根据内存地址/序列化信息获取NBT */
 export function getCompoundTag(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
+    info: number | NbtCompound | EventListenerData
 ): NbtCompound | undefined;
 
 /** 根据内存地址/序列化信息获取数字 */
 export function getLongLong(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number;
+    info: number | NbtCompound | EventListenerData
+): number;
 
 /** 根据内存地址/序列化信息获取数字 */
 export function getUnsignedLongLong(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number;
+    info: number | NbtCompound | EventListenerData
+): number;
 
 /** 根据内存地址/序列化信息获取数字 */
 export function getInt(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number;
+    info: number | NbtCompound | EventListenerData
+): number;
 
 /** 根据内存地址/序列化信息获取数字 */
 export function getUnsignedInt(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number;
+    info: number | NbtCompound | EventListenerData
+): number;
 
 /** 根据内存地址/序列化信息获取数字 */
 export function getShort(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number;
+    info: number | NbtCompound | EventListenerData
+): number;
 
 /** 根据内存地址/序列化信息获取数字 */
 export function getUnsignedShort(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number;
+    info: number | NbtCompound | EventListenerData
+): number;
 
 /** 根据内存地址/序列化信息获取数字 */
 export function getChar(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number;
+    info: number | NbtCompound | EventListenerData
+): number;
 
 /** 根据内存地址/序列化信息获取数字 */
 export function getUnsignedChar(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number;
+    info: number | NbtCompound | EventListenerData
+): number;
 
 /** 根据内存地址/序列化信息获取浮点数 */
 export function getFloat(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number;
+    info: number | NbtCompound | EventListenerData
+): number;
 
 /** 根据内存地址/序列化信息获取浮点数 */
 export function getDouble(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number;
+    info: number | NbtCompound | EventListenerData
+): number;
 
 /** 根据内存地址/序列化信息获取浮点数 */
 export function getLongDouble(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number;
+    info: number | NbtCompound | EventListenerData
+): number;
 
 /** 根据内存地址/序列化信息获取布尔值 */
 export function getBoolean(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
+    info: number | NbtCompound | EventListenerData
 ): Boolean;
 
 /** 根据内存地址/序列化信息获取字符串 */
 export function getString(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): String;
+    info: number | NbtCompound | EventListenerData
+): string;
 
 /** 根据内存地址/序列化信息获取指针原始地址 */
 export function getRawAddress(
     /** 地址/序列化信息 */
-    info: Number | NbtCompound | EventListenerData
-): Number | undefined;
+    info: number | NbtCompound | EventListenerData
+): number | undefined;
 
 /** 获取实体/玩家/物品/方块/方块实体/容器/NBT地址 */
 export function getAddress(
     target: Entity | Player | Item | Block | BlockEntity | Container | CompoundTag
-): Number | undefined;
+): number | undefined;
 
 declare let nbtProxyHandler: ProxyHandler<any>;
 
 interface NbtList {
-    toSNBT(index = -1): String;
-    forEach(callback: (value: any, index: Number, array: NbtList) => void): void;
-    filter(callback: (value: any, index: Number, array: NbtList) => Boolean): NbtList;
-    map(callback: (value: any, index: Number, array: NbtList) => any): NbtList;
-    find(callback: (value: any, index: Number, array: NbtList) => Boolean): any;
-    some(callback: (value: any, index: Number, array: NbtList) => Boolean): Boolean;
-    every(callback: (value: any, index: Number, array: NbtList) => Boolean): Boolean;
-    reduce<T = any>(callback: (accumulator: T, value: any, index: Number, array: NbtList) => any, initialValue?: T): T;
+    toSNBT(index = -1): string;
+    forEach(callback: (value: any, index: number, array: NbtList) => void): void;
+    filter(callback: (value: any, index: number, array: NbtList) => Boolean): NbtList;
+    map(callback: (value: any, index: number, array: NbtList) => any): NbtList;
+    find(callback: (value: any, index: number, array: NbtList) => Boolean): any;
+    some(callback: (value: any, index: number, array: NbtList) => Boolean): Boolean;
+    every(callback: (value: any, index: number, array: NbtList) => Boolean): Boolean;
+    reduce<T = any>(callback: (accumulator: T, value: any, index: number, array: NbtList) => any, initialValue?: T): T;
 }
 interface NbtByte {
-    toSNBT(): String;
+    toSNBT(): string;
 }
 interface NbtDouble {
-    toSNBT(): String;
+    toSNBT(): string;
 }
 interface NbtFloat {
-    toSNBT(): String;
+    toSNBT(): string;
 }
 interface NbtLong {
-    toSNBT(): String;
+    toSNBT(): string;
 }
 interface NbtInt {
-    toSNBT(): String;
+    toSNBT(): string;
 }
 interface NbtShort {
-    toSNBT(): String;
+    toSNBT(): string;
 }
 interface NbtString {
-    toSNBT(): String;
+    toSNBT(): string;
 }